refactor: extract buildChartData into chartDataBuilder.ts
Move the 230-line buildChartData method from CopilotTokenTracker into a
dedicated chartDataBuilder.ts module that accepts an explicit deps interface.

This follows the existing _functionName import pattern used throughout
extension.ts for statsHelpers, workspaceHelpers, tokenEstimation, etc.

Changes:
- Add vscode-extension/src/chartDataBuilder.ts with buildChartData() and
  ChartDataBuilderDeps interface; deps are injectable for testability
- Replace the 230-line method body in extension.ts with a thin 7-line wrapper
- Add optional now?: Date dep for deterministic tests
