+++ vscode-extension/src/extension.ts
@@ -151,6 +151,7 @@ import {
 } from './viewRegression';
 import { determineOnboardingAction } from './onboarding';
 import { addModelUsage, addEditorUsage, computeUtcDateRanges, aggregatePeriodStats, type SessionAggregateInput } from './statsHelpers';
+import { buildChartData as _buildChartData } from './chartDataBuilder';
 
 type LocalViewRegressionProbeResult = {
   pass: boolean;
@@ -8177,234 +8178,12 @@ ${hashtag}`;
   }
 
   private buildChartData(fullDailyStats: DailyTokenStats[]): ChartDataPayload {
-    const now = new Date();
-
-    const modelColors = [
-      { bg: "rgba(54, 162, 235, 0.6)", border: "rgba(54, 162, 235, 1)" },
-      { bg: "rgba(255, 99, 132, 0.6)", border: "rgba(255, 99, 132, 1)" },
-      { bg: "rgba(75, 192, 192, 0.6)", border: "rgba(75, 192, 192, 1)" },
-      { bg: "rgba(153, 102, 255, 0.6)", border: "rgba(153, 102, 255, 1)" },
-      { bg: "rgba(255, 159, 64, 0.6)", border: "rgba(255, 159, 64, 1)" },
-      { bg: "rgba(255, 205, 86, 0.6)", border: "rgba(255, 205, 86, 1)" },
-      { bg: "rgba(201, 203, 207, 0.6)", border: "rgba(201, 203, 207, 1)" },
-      { bg: "rgba(100, 181, 246, 0.6)", border: "rgba(100, 181, 246, 1)" },
-    ];
-
-    const fmtKey = (d: Date) =>
-      `${d.getFullYear()}-${String(d.getMonth() + 1).padStart(2, "0")}-${String(d.getDate()).padStart(2, "0")}`;
-
-    const emptyEntry = (date: string): DailyTokenStats => ({
-      date, tokens: 0, sessions: 0, interactions: 0,
-      modelUsage: {}, editorUsage: {}, repositoryUsage: {},
-    });
-
-    const mergeInto = (target: DailyTokenStats, src: DailyTokenStats) => {
-      target.tokens += src.tokens;
-      target.sessions += src.sessions;
-      target.interactions += src.interactions;
-      addModelUsage(target.modelUsage, src.modelUsage);
-      for (const [e, u] of Object.entries(src.editorUsage)) {
-        if (!target.editorUsage[e]) { target.editorUsage[e] = { tokens: 0, sessions: 0 }; }
-        target.editorUsage[e].tokens += u.tokens;
-        target.editorUsage[e].sessions += u.sessions;
-      }
-      for (const [r, u] of Object.entries(src.repositoryUsage)) {
-        if (!target.repositoryUsage[r]) { target.repositoryUsage[r] = { tokens: 0, sessions: 0 }; }
-        target.repositoryUsage[r].tokens += u.tokens;
-        target.repositoryUsage[r].sessions += u.sessions;
-      }
-    };
-
-    type BucketEntry = { label: string; key: string; stats: DailyTokenStats };
-
-    const buildPeriodData = (buckets: BucketEntry[]) => {
-      const entries = buckets.map(b => b.stats);
-      const labels = buckets.map(b => b.label);
-      const tokensData = entries.map(e => e.tokens);
-      const sessionsData = entries.map(e => e.sessions);
-
-      const allModels = new Set<string>();
-      entries.forEach(e => Object.keys(e.modelUsage).forEach(m => allModels.add(m)));
-
-      // Rank models by total tokens across the period; keep top 5, group the rest
-      const modelTotals = new Map<string, number>();
-      for (const model of allModels) {
-        const total = entries.reduce((sum, e) => {
-          const u = e.modelUsage[model];
-          return sum + (u ? u.inputTokens + u.outputTokens : 0);
-        }, 0);
-        modelTotals.set(model, total);
-      }
-      const sortedModels = Array.from(allModels).sort((a, b) => (modelTotals.get(b) || 0) - (modelTotals.get(a) || 0));
-      const topModels = sortedModels.slice(0, 5);
-      const otherModels = sortedModels.slice(5);
-
-      const modelDatasets = topModels.map((model, idx) => {
-        const color = modelColors[idx % modelColors.length];
-        return {
-          label: getModelDisplayName(model),
-          data: entries.map(e => { const u = e.modelUsage[model]; return u ? u.inputTokens + u.outputTokens : 0; }),
-          backgroundColor: color.bg, borderColor: color.border, borderWidth: 1,
-        };
-      });
-      if (otherModels.length > 0) {
-        modelDatasets.push({
-          label: 'Other models',
-          data: entries.map(e => otherModels.reduce((sum, m) => {
-            const u = e.modelUsage[m];
-            return sum + (u ? u.inputTokens + u.outputTokens : 0);
-          }, 0)),
-          backgroundColor: 'rgba(150, 150, 150, 0.5)',
-          borderColor: 'rgba(150, 150, 150, 0.8)',
-          borderWidth: 1,
-        });
-      }
-
-      const allEditors = new Set<string>();
-      entries.forEach(e => Object.keys(e.editorUsage).forEach(ed => allEditors.add(ed)));
-      const editorDatasets = Array.from(allEditors).map((editor, idx) => {
-        const color = modelColors[idx % modelColors.length];
-        return {
-          label: editor,
-          data: entries.map(e => e.editorUsage[editor]?.tokens || 0),
-          backgroundColor: color.bg, borderColor: color.border, borderWidth: 1,
-        };
-      });
-
-      const allRepos = new Set<string>();
-      entries.forEach(e => Object.keys(e.repositoryUsage)
-        .filter(r => r !== 'Unknown')
-        .forEach(r => allRepos.add(r)));
-      const repositoryDatasets = Array.from(allRepos).map((repo, idx) => {
-        const color = modelColors[idx % modelColors.length];
-        return {
-          label: this.getRepoDisplayName(repo),
-          fullRepo: repo,
-          data: entries.map(e => e.repositoryUsage[repo]?.tokens || 0),
-          backgroundColor: color.bg, borderColor: color.border, borderWidth: 1,
-        };
-      });
-
-      const totalTokens = tokensData.reduce((a, b) => a + b, 0);
-      const totalSessions = sessionsData.reduce((a, b) => a + b, 0);
-      const periodCount = buckets.length;
-
-      const costData = entries.map(e => this.calculateEstimatedCost(e.modelUsage, 'copilot'));
-      const totalCost = costData.reduce((a, b) => a + b, 0);
-
-      return {
-        labels, tokensData, sessionsData, modelDatasets, editorDatasets, repositoryDatasets,
-        periodCount, totalTokens, totalSessions,
-        avgPerPeriod: periodCount > 0 ? Math.round(totalTokens / periodCount) : 0,
-        costData,
-        totalCost,
-        avgCostPerPeriod: periodCount > 0 ? totalCost / periodCount : 0,
-      };
-    };
-
-    // ── Daily period: last 30 days with zero-fill ─────────────────────
-    const thirtyDaysAgo = new Date(now.getFullYear(), now.getMonth(), now.getDate() - 30);
-    const thirtyDaysAgoStr = fmtKey(thirtyDaysAgo);
-    const todayStr = fmtKey(now);
-    const dailyBucketMap = new Map<string, BucketEntry>();
-    for (let cursor = new Date(thirtyDaysAgo); cursor <= now; cursor.setDate(cursor.getDate() + 1)) {
-      const key = fmtKey(new Date(cursor));
-      dailyBucketMap.set(key, { key, label: key, stats: emptyEntry(key) });
-    }
-    for (const day of fullDailyStats) {
-      if (day.date >= thirtyDaysAgoStr && day.date <= todayStr) {
-        const bucket = dailyBucketMap.get(day.date);
-        if (bucket) { mergeInto(bucket.stats, day); }
-      }
-    }
-    const dailyBuckets = Array.from(dailyBucketMap.values()).sort((a, b) => a.key.localeCompare(b.key));
-    const dailyPeriod = buildPeriodData(dailyBuckets);
-
-    // ── Weekly period: last 6 calendar weeks with zero-fill ──────────
-    const getMondayOfWeek = (d: Date): Date => {
-      const copy = new Date(d); copy.setHours(0, 0, 0, 0);
-      const day = copy.getDay();
-      copy.setDate(copy.getDate() - (day === 0 ? 6 : day - 1));
-      return copy;
-    };
-    const fmtWeekLabel = (monday: Date): string => {
-      const sunday = new Date(monday); sunday.setDate(monday.getDate() + 6);
-      if (monday.getMonth() === sunday.getMonth()) {
-        return `${monday.toLocaleDateString("en-US", { month: "short" })} ${monday.getDate()}–${sunday.getDate()}`;
-      }
-      return `${monday.toLocaleDateString("en-US", { month: "short", day: "numeric" })}–${sunday.toLocaleDateString("en-US", { month: "short", day: "numeric" })}`;
-    };
-    const thisMonday = getMondayOfWeek(now);
-    const weekBucketMap = new Map<string, BucketEntry>();
-    for (let w = 5; w >= 0; w--) {
-      const monday = new Date(thisMonday); monday.setDate(thisMonday.getDate() - w * 7);
-      const key = fmtKey(monday);
-      weekBucketMap.set(key, { key, label: fmtWeekLabel(monday), stats: emptyEntry(key) });
-    }
-    for (const day of fullDailyStats) {
-      const monday = getMondayOfWeek(new Date(day.date + "T00:00:00"));
-      const bucket = weekBucketMap.get(fmtKey(monday));
-      if (bucket) { mergeInto(bucket.stats, day); }
-    }
-    const weeklyBuckets = Array.from(weekBucketMap.values()).sort((a, b) => a.key.localeCompare(b.key));
-    const weeklyPeriod = buildPeriodData(weeklyBuckets);
-
-    // ── Monthly period: last 12 calendar months with zero-fill ───────
-    const monthBucketMap = new Map<string, BucketEntry>();
-    for (let m = 11; m >= 0; m--) {
-      const monthDate = new Date(now.getFullYear(), now.getMonth() - m, 1);
-      const key = `${monthDate.getFullYear()}-${String(monthDate.getMonth() + 1).padStart(2, "0")}`;
-      const label = monthDate.toLocaleDateString("en-US", { month: "short", year: "numeric" });
-      monthBucketMap.set(key, { key, label, stats: emptyEntry(key) });
-    }
-    for (const day of fullDailyStats) {
-      const monthKey = day.date.slice(0, 7);
-      const bucket = monthBucketMap.get(monthKey);
-      if (bucket) { mergeInto(bucket.stats, day); }
-    }
-    const monthlyBuckets = Array.from(monthBucketMap.values()).sort((a, b) => a.key.localeCompare(b.key));
-    const monthlyPeriod = buildPeriodData(monthlyBuckets);
-
-    // ── Summary totals from the daily period (last 30 days) ──────────
-    const editorTotalsMap: Record<string, number> = {};
-    dailyBuckets.forEach(b => {
-      Object.entries(b.stats.editorUsage).forEach(([editor, usage]) => {
-        editorTotalsMap[editor] = (editorTotalsMap[editor] || 0) + usage.tokens;
-      });
-    });
-    const repositoryTotalsMap: Record<string, number> = {};
-    dailyBuckets.forEach(b => {
-      Object.entries(b.stats.repositoryUsage)
-        .filter(([repo]) => repo !== 'Unknown')
-        .forEach(([repo, usage]) => {
-        const displayName = this.getRepoDisplayName(repo);
-        repositoryTotalsMap[displayName] = (repositoryTotalsMap[displayName] || 0) + usage.tokens;
-      });
-    });
-
-    return {
-      // Backward-compat flat fields (daily period)
-      labels: dailyPeriod.labels,
-      tokensData: dailyPeriod.tokensData,
-      sessionsData: dailyPeriod.sessionsData,
-      modelDatasets: dailyPeriod.modelDatasets,
-      editorDatasets: dailyPeriod.editorDatasets,
-      repositoryDatasets: dailyPeriod.repositoryDatasets,
-      editorTotalsMap,
-      repositoryTotalsMap,
-      dailyCount: dailyPeriod.periodCount,
-      totalTokens: dailyPeriod.totalTokens,
-      avgTokensPerDay: dailyPeriod.avgPerPeriod,
-      totalSessions: dailyPeriod.totalSessions,
-      lastUpdated: new Date().toISOString(),
+    return _buildChartData(fullDailyStats, {
+      getRepoDisplayName: _getRepoDisplayName,
+      calculateEstimatedCost: (modelUsage, pricingSource) => _calculateEstimatedCost(modelUsage, this.modelPricing, pricingSource),
       backendConfigured: this.isBackendConfigured(),
       compactNumbers: this.getCompactNumbersSetting(),
-      periods: {
-        day: dailyPeriod,
-        week: weeklyPeriod,
-        month: monthlyPeriod,
-      },
-    };
+    });
   }
 
   private getChartHtml(
